requirements.txt
apache-beam[gcp]
Faker
faker-schema
google-cloud-bigquery
google-cloud-bigquery-storage
mock
numpy
pandas